+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - possibility to set a user-defined MQTT client ID for both, agent and client [#56]
+- possibility to set a user-defined VRPC client identity
 
 ### Fixed